cast from 'LPBYTE {aka unsigned char*}' to 'uint32 {aka unsigned int}' loses precision but not any more, as reverted to uint64

+++ neo/sys/win32/win_cpu.cpp
@@ -130,7 +130,7 @@ double Sys_ClockTicksPerSecond() {
 			}
 			RegCloseKey( hKey );
 			if ( ret == ERROR_SUCCESS ) {
-				ticks = (double) ((uint32)ProcSpeed) * 1000000;
+				ticks = (double) ((uint64)ProcSpeed) * 1000000;
 			}
 		}
 	}
